Updated uint types

+++ src/LiquidityPool.sol
@@ -132,8 +132,6 @@ contract LiquidityPool is Initializable, OwnableUpgradeable, UUPSUpgradeable, IL
         totalValueInLp -= uint128(_amount);
         if (_amount > type(uint128).max || _amount == 0 || share == 0) revert InvalidAmount();
 
-        //Potentially we could do this if statement offline?
-        //emit FundsWithdrawn(msg.sender, _amount);
         if(msg.sender == address(membershipManager)) {
             emit FundsWithdrawn(SourceOfFunds.ETHER_FAN, _amount);
         } else {
@@ -192,7 +190,7 @@ contract LiquidityPool is Initializable, OwnableUpgradeable, UUPSUpgradeable, IL
         (uint256 firstIndex, uint128 lastIndex, uint128 lastIndexNumOfValidators) = dutyForWeek();
         _isAssigned(firstIndex, lastIndex, _index);
         require(msg.sender == bnftHolders[_index].holder, "Incorrect Caller");
-        require(bnftHolders[_index].timestamp < _getCurrentSchedulingStartTimestamp(), "Already deposited");
+        require(bnftHolders[_index].timestamp < uint32(_getCurrentSchedulingStartTimestamp()), "Already deposited");
 
         uint256 numberOfValidatorsToSpin = max_validators_per_owner;
         if(_index == lastIndex) {
@@ -209,7 +207,7 @@ contract LiquidityPool is Initializable, OwnableUpgradeable, UUPSUpgradeable, IL
         totalValueInLp -= uint128(amountFromLp);
         numPendingDeposits += uint32(numberOfValidatorsToSpin);
 
-        bnftHolders[_index].timestamp = block.timestamp;
+        bnftHolders[_index].timestamp = uint32(block.timestamp);
 
         uint256[] memory newValidators = stakingManager.batchDepositWithBidIds{value: 32 ether * numberOfValidatorsToSpin}(_candidateBidIds, _merkleProof, msg.sender);
 
@@ -401,9 +399,9 @@ contract LiquidityPool is Initializable, OwnableUpgradeable, UUPSUpgradeable, IL
         emit UpdatedSchedulingPeriod(_schedulingPeriodInSeconds);
     }
 
-    function numberOfActiveSlots(BnftHolder[] memory _localBnftHoldersArray) public view returns (uint256 numberOfActiveSlots) {
-        numberOfActiveSlots = uint128(_localBnftHoldersArray.length);
-        if(holdersUpdate.timestamp > uint128(_getCurrentSchedulingStartTimestamp())) {
+    function numberOfActiveSlots(BnftHolder[] memory _localBnftHoldersArray) public view returns (uint32 numberOfActiveSlots) {
+        numberOfActiveSlots = uint32(_localBnftHoldersArray.length);
+        if(holdersUpdate.timestamp > uint32(_getCurrentSchedulingStartTimestamp())) {
             numberOfActiveSlots = holdersUpdate.startOfSlotNumOwners;
         }
     }
@@ -413,10 +411,10 @@ contract LiquidityPool is Initializable, OwnableUpgradeable, UUPSUpgradeable, IL
     //--------------------------------------------------------------------------------------
 
     function _checkHoldersUpdateStatus() internal {
-        if(holdersUpdate.timestamp < uint128(_getCurrentSchedulingStartTimestamp())) {
-            holdersUpdate.startOfSlotNumOwners = uint128(bnftHolders.length);
+        if(holdersUpdate.timestamp < uint32(_getCurrentSchedulingStartTimestamp())) {
+            holdersUpdate.startOfSlotNumOwners = uint32(bnftHolders.length);
         }
-        holdersUpdate.timestamp = uint128(block.timestamp);
+        holdersUpdate.timestamp = uint32(block.timestamp);
     }
 
     function _getCurrentSchedulingStartTimestamp() internal view returns (uint256) {
@@ -436,7 +434,7 @@ contract LiquidityPool is Initializable, OwnableUpgradeable, UUPSUpgradeable, IL
     }
 
     function _fetchLastIndex(uint128 _size, uint256 _index, BnftHolder[] memory _localBnftHoldersArray) internal returns (uint128 lastIndex){
-        uint256 numSlots = numberOfActiveSlots(_localBnftHoldersArray);
+        uint32 numSlots = numberOfActiveSlots(_localBnftHoldersArray);
         uint128 tempLastIndex = uint128(_index) + _size - 1;
         lastIndex = (tempLastIndex + uint128(numSlots)) % uint128(numSlots);
     }
